Association fails with certain subnet names
Subnets that are name `subnet-XYZ-environment-x-zone-y` are not being associated with a newly created or updated NACL. Currently the code checks for `startswith('subnet-')` and when matching assumes it is a subnet-id. I am no python coder (nor much of a coder at all) but when the order is reversed (first checking for name and then for ID), this would fix it. Until then I would recommend a little amendment in the documentation.